Group `shared.ts` and `util.ts`
