+++ libs/@blockprotocol/graph/src/shared/stdlib/subgraph/edge/link-entity.ts
@@ -1,4 +1,4 @@
-import { typedEntries } from "../../../../shared.js";
+import { typedEntries } from "../../../../util";
 import {
   Entity,
   EntityId,
