+++ libs/@blockprotocol/graph/src/shared/stdlib/subgraph/element/data-type.ts
@@ -5,7 +5,7 @@ import {
   VersionedUri,
 } from "@blockprotocol/type-system/slim";
 
-import { typedValues } from "../../../../shared.js";
+import { typedValues } from "../../../../util";
 import { DataTypeWithMetadata } from "../../../types/ontology/data-type.js";
 import { OntologyTypeVertexId, Subgraph } from "../../../types/subgraph.js";
 import { isDataTypeVertex } from "../../../types/subgraph/vertices.js";
