+++ libs/@blockprotocol/graph/src/shared/stdlib/subgraph/element/property-type.ts
@@ -5,7 +5,7 @@ import {
   VersionedUri,
 } from "@blockprotocol/type-system/slim";
 
-import { typedValues } from "../../../../shared.js";
+import { typedValues } from "../../../../util";
 import { PropertyTypeWithMetadata } from "../../../types/ontology/property-type.js";
 import { OntologyTypeVertexId, Subgraph } from "../../../types/subgraph.js";
 import { isPropertyTypeVertex } from "../../../types/subgraph/vertices.js";
